+++ packages/toolboxes/substrate/src/toolbox/baseSubstrateToolbox.ts
@@ -94,7 +94,7 @@ const transfer = async (
   return tx?.toString();
 };
 
-const estimateGasFee = async (
+const estimateTransactionFee = async (
   api: ApiPromise,
   signer: IKeyringPair | Signer,
   gasAsset: AssetValue,
@@ -179,8 +179,8 @@ export const BaseSubstrateToolbox = ({
   getBalance: (address: string) => getBalance(api, gasAsset, address),
   validateAddress: (address: string) => validateAddress(address, network.prefix),
   transfer: (params: SubstrateTransferParams) => transfer(api, signer, params),
-  estimateGasFee: (params: SubstrateTransferParams) =>
-    estimateGasFee(api, signer, gasAsset, params),
+  estimateTransactionFee: (params: SubstrateTransferParams) =>
+    estimateTransactionFee(api, signer, gasAsset, params),
   sign: (tx: SubmittableExtrinsic<"promise">) => {
     if (isKeyringPair(signer)) {
       return sign(signer, tx);