+++ packages/swapkit/chainflip/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @swapkit/chainflip
 
+## 1.0.0-rc.68
+
+### Patch Changes
+
+- Updated dependencies [[`ab5c3b3`](https://github.com/thorswap/SwapKit/commit/ab5c3b3e2b8987c2496c8d375e8c4aaa8ffb5f4d)]:
+  - @swapkit/toolbox-substrate@1.0.0-rc.47
+  - @swapkit/helpers@1.0.0-rc.112
+  - @swapkit/toolbox-evm@1.0.0-rc.119
+  - @swapkit/api@1.0.0-rc.80
+
 ## 1.0.0-rc.67
 
 ### Patch Changes