+++ packages/wallets/xdefi/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @swapkit/wallet-xdefi
 
+## 1.0.0-rc.160
+
+### Patch Changes
+
+- Updated dependencies [[`ab5c3b3`](https://github.com/thorswap/SwapKit/commit/ab5c3b3e2b8987c2496c8d375e8c4aaa8ffb5f4d)]:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.139
+  - @swapkit/helpers@1.0.0-rc.112
+  - @swapkit/toolbox-utxo@1.0.0-rc.123
+  - @swapkit/toolbox-evm@1.0.0-rc.119
+
 ## 1.0.0-rc.159
 
 ### Patch Changes